Add test to see if exhibit is set for this view.
Notice entries in /culprod/omeka/logs/error_log stating:
BEGIN_BLURB>>>>>>
Omeka fatal error: exception 'Omeka_View_Exception' with message
'A current exhibit variable has not been set to this view.'
<<<<<<END_BLURB
The above is just the beginning of the entry, the entry contains
a stack trace. Part of the trace contained the following:
BEGIN_BLURB>>>>>
/cul/cul0/lito/vmounts/haydn/var-www/omeka/ssl-html-222/plugins/
ExhibitBuilder/helpers/ExhibitFunctions.php(76): get_current_record('exhibit')
<<<<<<END_BLURB.
Using the timestamp on the error, I found the associated entry in access_log:
BEGIN_BLURB>>>>>>
66.249.67.82 - - [07/Jun/2015:05:42:13 -0400]
"GET /exhibits/show/lit_hum/shakespeare/item/6182 HTTP/1.1"
500 1369 "-" "Mozilla/5.0 (compatible; Googlebot/2.1; +http://www.google.com/bot.html)"
<<<<<<END_BLURB
This is not a valid URL -- it is not the correct URL to get to the exhibit item page
for item 6182. Not sure how the google bot is generating this URL.
Added a test to make sure the exhibit variable is set for this view.